# docs/swagger.yaml
basePath: /v1
definitions:
  auth.RegisterRequest:
    properties:
      birth_date:
        example: "1998-05-14T00:00:00Z"
        type: string
      email:
        example: john@example.com
        type: string
      first_name:
        example: john
        maxLength: 100
        minLength: 3
        type: string
      last_name:
        example: doe
        maxLength: 100
        minLength: 3
        type: string
      password:
        minLength: 8
        type: string
      username:
        example: john
        maxLength: 100
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
  auth.UserResponse:
    properties:
      birth_date:
        example: "1998-05-14T00:00:00Z"
        type: string
      email:
        example: john@example.com
        type: string
      first_name:
        example: john
        type: string
      last_name:
        example: doe
        type: string
      username:
        example: john
        type: string
      xid:
        example: 9m4e2mr0ui3e8a215n4g
        type: string
    type: object
  httpx.ErrorResponse:
    properties:
      message:
        example: internal server error
        type: string
    type: object
  users.CreateUserRequest:
    properties:
      birth_date:
        example: "1998-05-14T00:00:00Z"
        type: string
      email:
        example: john@example.com
        type: string
      first_name:
        example: john
        maxLength: 100
        minLength: 3
        type: string
      last_name:
        example: doe
        maxLength: 100
        minLength: 3
        type: string
      username:
        example: john
        maxLength: 100
        minLength: 3
        type: string
    required:
    - email
    - username
    type: object
  users.UserResponse:
    properties:
      birth_date:
        example: "1998-05-14T00:00:00Z"
        type: string
      email:
        example: john@example.com
        type: string
      first_name:
        example: john
        type: string
      last_name:
        example: doe
        type: string
      username:
        example: john
        type: string
      xid:
        example: 9m4e2mr0ui3e8a215n4g
        type: string
    type: object
info:
  contact: {}
  description: A production-ready Go API template.
  title: Go API Template
  version: "1.0"
paths:
  /admin/users:
    get:
      description: Get all users
      parameters:
      - default: 1
        description: Page number
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/users.UserResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: List users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user
      parameters:
      - description: User payload
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/users.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/users.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Create user
      tags:
      - users
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user
      produces:
      - application/json
      responses: {}
      summary: Login
      tags:
      - auth
  /auth/logout:
    post:
      description: Logout the current user
      produces:
      - application/json
      responses: {}
      summary: Logout
      tags:
      - auth
  /auth/me:
    get:
      description: Get the currently authenticated user
      produces:
      - application/json
      responses: {}
      summary: Current user
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user account
      parameters:
      - description: Registration payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/auth.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Register
      tags:
      - auth
swagger: "2.0"